use accent colour for mania snap colouring
This commit switches to using AccentColour (instead of the previous
Colour) for snap colouring in osu!mania. Snap colouring is now also
applied to all mania DHOs. This improves the look and feel of Argon
with snap colouring turned on from the original broken-ish state.

Snap colouring in Triangle is now very subtle and only clearly visible
on hold notes, while it is effectively removed from legacy skins.